pyproject.toml
[project]
name = "crossword-space"
version = "0.1.0"
description = "Latent Manifold Learning for Italian Crosswords and Beyond"
readme = "README.md"
requires-python = ">=3.10"
license = "CC-BY-4.0"
authors = [
    { name = "Cristiano Ciaccio" },
    { name = "Gabriele Sarti" },
    { name = "Alessio Miaschi" },
    { name = "Felice Dell'Orletta" },
]

dependencies = [
    "torch>=2.0",
    "transformers>=4.30",
    "datasets>=2.14",
    "pandas>=2.0",
    "numpy>=1.24",
    "faiss-cpu>=1.7",
    "tqdm>=4.60",
    "tensorboard>=2.14",
    "tabulate>=0.9",
    "z3-solver>=4.12",
    "matplotlib>=3.7",
    "python-Levenshtein>=0.21",
    "rank-bm25>=0.2",
]

[tool.setuptools.packages.find]
include = ["crossword_space*"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.11",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
